[WIP]refactor gui
