+++ doc/copyright.txt
@@ -1,0 +1,4 @@
+【技术支持】
+联系开发者：18372124178
+电子邮箱：wankj@cug.edu.cn
+更新日期：2024-03-04
