+++ doc/user_manual.txt
@@ -1,0 +1,87 @@
+======================== 用户手册 ========================
+欢迎使用3D航线规划仿真系统
+
+【程序概述】
+本程序提供三维场景下的无人机航线规划与飞行仿真功能，支持：
+- OBJ格式3D模型加载与展示
+- 二维地图联动显示
+- 交互式航线规划
+- 三维飞行仿真
+- 多视角观察
+
+【界面布局】
+1. 左侧控制面板
+   - 视图切换：3D视图/2D地图/重置视角
+   - 航线规划：设置基准高度/飞行高度/航带宽度
+   - 仿真控制：启动/暂停/继续/返航
+2. 右侧属性面板
+   - 文件树：显示加载的3D模型目录结构
+   - 工具箱：快速访问常用工具
+3. 中央视图区
+   - 3D视图：支持鼠标拖拽旋转/滚轮缩放
+   - 2D地图：集成OpenStreetMap在线地图
+
+【核心功能指南】
+▶ 加载3D模型：
+   1. 点击菜单栏 Project -> load 3D file
+   2. 选择包含.obj和.jpg的文件夹
+   3. 模型将自动加载到3D视图
+
+▶ 创建航线：
+   1. 在控制面板设置基准高度（地面高度）
+   2. 点击'Create Route'进入规划模式
+   3. 操作步骤：
+      a) 点击'Set Home Point'设置起飞点
+      b) 点击'Add Control Point'添加控制点
+      c) 调整航带宽度（相邻航线间距）
+      d) 点击'Generate Route'生成最终航线
+   4. 右键点击控制点可进行删除操作
+
+▶ 飞行仿真：
+   1. 设置飞行速度（建议5-15 m/s）
+   2. 操作按钮：
+      - Start：从Home点开始仿真
+      - Pause：暂停仿真
+      - Resume：继续飞行
+      - Return Home：立即返航
+      - Stop：终止仿真
+   3. 按空格键切换摄像机跟随模式
+
+▶ 视角操作：
+   - 鼠标左键拖拽：旋转视角
+   - 鼠标滚轮：缩放视图
+   - 方向键：平移视图
+   - Reset按钮：恢复默认视角
+
+【参数说明】
+▶ 飞行参数：
+   - 基准高度：模型所在平面高度（单位：米）
+   - 飞行高度：相对基准面的垂直高度（±1000m）
+   - 航带宽度：相邻航线水平间距（1-300m）
+   - 飞行速度：仿真速度（1-50m/s）
+
+▶ 环境参数：
+   - 温度：模拟环境温度（-20°C~45°C）
+     影响电池效率（低温降低续航）
+   - 气压：大气压强（950-1050hPa）
+     用于高度计校准参考
+   - 天气状况：模拟气象条件
+     ■ 晴天：正常能见度
+     ■ 雨天：降低最大飞行速度20%
+     ■ 雪天：缩短有效控制距离30%
+     ■ 雾天：开启防撞雷达
+
+▶ 系统参数：
+   - 模型缩放：默认1:1比例尺
+   - 纹理精度：2048x2048像素
+   - 刷新频率：60Hz（垂直同步）
+
+【常见问题】
+Q1: 控制点添加不成功？
+A: 请确保：1) 已进入规划模式 2) 鼠标点击位置在模型表面
+
+Q2: 生成的航线不符合预期？
+A: 检查控制点是否形成有效凸包，建议至少设置3个控制点
+
+Q3: 3D模型显示异常？
+A: 确认模型文件包含配套的.jpg纹理文件，且尺寸为2的幂次方
