+++ src/MainWindow.h
@@ -9,14 +9,10 @@ Date:2025.1.6
 #include "gui/Canvas.h"
 #include "gui/Menu.h"
 #include "gui/DockWidget.h"
-#include "log/QgisDebug.h"
-#include "ui_MainWindow.h"
 #include <QAction>
 #include <QApplication>
 #include <QKeyEvent>
-
-//#include <JoystickWidget.h>
-//#include <qgamepad.h>
+#include <QMainWindow>
 
 class MainWindow : public QMainWindow {
   Q_OBJECT
@@ -27,80 +23,18 @@ public:
     return instance;
   }
   ~MainWindow();
-
-private:
-  MainWindow(QWidget *parent = nullptr);
   MainWindow(const MainWindow&) = delete;
   MainWindow& operator=(const MainWindow&) = delete;
 
-  // void setTianDiTuMap(double lat, double lon, int zoom);
-  QLabel *mpImageLabel;
-private slots:
-
 private:
-  void createMenu();
-  void createJoyDockWidgets();
-  void createLeftDockWidget();
-  void createRightDockWidget();
-  void createSlots();
+  MainWindow(QWidget *parent = nullptr);
   Canvas *mpCanvas;
   MenuBar *mpMenuBar;
   LeftDockWidget *mpLeftDockWidget;
   RightDockWidget *mpRightDockWidget;
-  QTreeWidget *mpFileTreeWidget;
-  QPushButton *mpBtnReset;
-  QPushButton *mpBtnSwitchTo3D;
-  QPushButton *mpBtnSwitchTo2D;
-  void onSelectDirectoryClicked();
-  void loadDirectoryFiles(const QString &path);
-  void loadDirectoryLevel(QTreeWidgetItem *parentItem, const QString &path,
-                          int level, int maxLevel);
-  void onTreeItemExpanded(QTreeWidgetItem *item);
-  void onTreeItemDoubleClicked(QTreeWidgetItem *item, int column);
-  QString getItemFullPath(QTreeWidgetItem *item);
+  void createSlots();
   void initWindowStatus();
-  template <typename Tp> // Tp is the pointer type
-  Tp safeFindChild(const QString &name) {
-    Tp pWidget = this->findChild<Tp>(name);
-    if (pWidget == nullptr) {
-      logMessage("findChild: " + name + " not found",
-                 Qgis::MessageLevel::Critical);
-      return dynamic_cast<Tp>(
-          ws::WindowManager::getInstance().getDefaultObject());
-    }
-    return pWidget;
-  }
-
-private slots:
-  void queryFlightParameters();
-  void refreshBasicData();
-  void loadModel(const QString &objFilePath);
-
-private:
-  QLabel *m_pFlightParamsDisplay;
-  QLabel *m_pWeatherLabel;
-  QLabel *m_pTemperatureLabel;
-  QLabel *m_pPressureLabel;
-  
-private slots:
-  void showFlightParamsDialog();
-  void showEnvironmentalParamsDialog();
   void showUserManual();
-  void Unrealized();
-
-private:
-  QPushButton *m_btnManualMode;
-  QPushButton *m_btnAutoMode;
-
-  // private slots:
-  //   void handleJoystickMove(float dx, float dy);
-  //   void switchToManualMode();
-  //   void switchToAutoMode();
-  //
-  // private:
-  //   QGamepad *m_gamepad = nullptr;
-  //   JoystickWidget *m_leftJoystick = nullptr;
-  //   JoystickWidget *m_rightJoystick = nullptr;
 };
 
 #endif // MAINWINDOW_H
